set priority for view docs button, move some command test to the correct folder
